feat(main.rs): implement resolve transaction logic
This commit introduces the logic for resolving disputed transactions.
When a 'resolve' action is processed, the system checks if there is
a corresponding 'dispute' action for the same transaction. If both
actions exist, the held amount is moved back to the available amount.

fix(main.rs): change actions storage to nested hashmap

The actions hashmap is changed to a nested hashmap to store
actions by client_id and tx_id. This allows to store multiple
actions for the same transaction.

feat(main.rs): add debug derive to TxAction enum

This commit adds the Debug derive macro to the TxAction enum to
allow for easier debugging.